+++ isy-security/src/test/java/de/bund/bva/isyfact/security/oauth2/client/AuthentifizierungsmanagerTest.java
@@ -1,8 +1,16 @@
 package de.bund.bva.isyfact.security.oauth2.client;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.Mockito.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -23,6 +31,7 @@ import org.springframework.security.oauth2.core.AuthorizationGrantType;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.TestPropertySource;
 
 import de.bund.bva.isyfact.security.AbstractOidcProviderTest;
 import de.bund.bva.isyfact.security.autoconfigure.IsyOAuth2ClientAutoConfiguration;
@@ -42,6 +51,10 @@ import de.bund.bva.isyfact.security.oauth2.client.authentication.token.PasswordC
  */
 @ActiveProfiles("test-clients")
 @SpringBootTest
+@TestPropertySource(properties = {
+        "isy.security.cache.ttl=300",
+        "isy.security.cache.maxelements=100"
+})
 public class AuthentifizierungsmanagerTest extends AbstractOidcProviderTest {
 
     @MockBean
@@ -67,12 +80,13 @@ public class AuthentifizierungsmanagerTest extends AbstractOidcProviderTest {
         SecurityContextHolder.getContext().setAuthentication(null);
 
         mockJwt = mock(JwtAuthenticationToken.class);
+        JwtAuthenticationToken secondMockJwt = mock(JwtAuthenticationToken.class);;
 
         when(clientCredentialsAuthorizedClientAuthenticationProvider.supports(any())).thenCallRealMethod();
         when(clientCredentialsAuthorizedClientAuthenticationProvider.authenticate(any(Authentication.class))).thenReturn(mockJwt);
 
         when(passwordClientRegistrationAuthenticationProvider.supports(any())).thenCallRealMethod();
-        when(passwordClientRegistrationAuthenticationProvider.authenticate(any(Authentication.class))).thenReturn(mockJwt);
+        when(passwordClientRegistrationAuthenticationProvider.authenticate(any(Authentication.class))).thenReturn(mockJwt, secondMockJwt);
 
         when(clientCredentialsClientRegistrationAuthenticationProvider.supports(any())).thenCallRealMethod();
         when(clientCredentialsClientRegistrationAuthenticationProvider.authenticate(any(Authentication.class))).thenReturn(mockJwt);
@@ -294,4 +308,75 @@ public class AuthentifizierungsmanagerTest extends AbstractOidcProviderTest {
         assertThat(exception).hasMessageContainingAll("unsupported-grant-type", "not supported");
     }
 
+    @Test
+    public void testCacheHitOnSecondAuthentication() {
+
+        // override value from @BeforeEach otherwise the authentication.isAuthenticated returns always false
+        when(mockJwt.isAuthenticated()).thenReturn(true);
+
+        authentifizierungsmanager.authentifiziereSystem(getIssuer(), "testid", "testsecret", "testuser", "testpw");
+        verify(passwordClientRegistrationAuthenticationProvider, times(1)).authenticate(any());
+        SecurityContextHolder.clearContext();
+
+        authentifizierungsmanager.authentifiziereSystem(getIssuer(), "testid", "testsecret", "testuser", "testpw");
+        verify(passwordClientRegistrationAuthenticationProvider, times(1)).authenticate(any());
+        assertNotNull(SecurityContextHolder.getContext().getAuthentication());
+
+        // clear context and cache so no other tests are affected
+        authentifizierungsmanager.clearCache();
+        SecurityContextHolder.clearContext();
+    }
+
+    @Test
+    public void testCacheWithSameUsernameButDifferentPassword() {
+        // First authentication attempt with credentials from testid1
+        // Provider is called
+        authentifizierungsmanager.authentifiziereSystem(getIssuer(), "testid", "testsecret", "testuser", "testpw");
+        verify(passwordClientRegistrationAuthenticationProvider, times(1)).authenticate(any());
+        Authentication firstAuthentication = SecurityContextHolder.getContext().getAuthentication();
+        SecurityContextHolder.clearContext();
+
+        // Second authentication attempt with credentials from testid1, but with different user password
+        // provider is called once again because the first stored cashedKey is different
+        authentifizierungsmanager.authentifiziereSystem(getIssuer(), "testid", "otherpw", "testuser", "otherpw");
+        verify(passwordClientRegistrationAuthenticationProvider, times(2)).authenticate(any());
+        Authentication secondAuthentication = SecurityContextHolder.getContext().getAuthentication();
+        assertNotNull(secondAuthentication);
+        assertNotEquals(firstAuthentication, secondAuthentication);
+
+        // clear context and cache so no other tests are affected
+        authentifizierungsmanager.clearCache();
+        SecurityContextHolder.clearContext();
+    }
+
+    @Test
+    public void testNoCachingForClientCredentialsRegistrationId() {
+        authentifizierungsmanager.authentifiziere("cc-client");
+        verify(clientCredentialsAuthorizedClientAuthenticationProvider, times(1)).authenticate(any());
+
+        SecurityContextHolder.clearContext();
+        authentifizierungsmanager.authentifiziere("cc-client");
+
+        verify(clientCredentialsAuthorizedClientAuthenticationProvider, times(2)).authenticate(any());
+
+        // clear context and cache so no other tests are affected
+        authentifizierungsmanager.clearCache();
+        SecurityContextHolder.clearContext();
+    }
+
+    @Test
+    public void testCacheClearedAfterClearCache() {
+        authentifizierungsmanager.authentifiziereClient(getIssuer(), "testid", "testsecret");
+        verify(clientCredentialsClientRegistrationAuthenticationProvider, times(1)).authenticate(any());
+
+        authentifizierungsmanager.clearCache();
+        SecurityContextHolder.clearContext();
+        authentifizierungsmanager.authentifiziereClient(getIssuer(), "testid", "testsecret");
+
+        verify(clientCredentialsClientRegistrationAuthenticationProvider, times(2)).authenticate(any());
+
+        // clear context and cache so no other tests are affected
+        authentifizierungsmanager.clearCache();
+        SecurityContextHolder.clearContext();
+    }
 }